whisper/requirements.txt
openai-whisper
ffmpeg
torch
torchvision 
torchaudio
spleeter
moviepy
Pillow
numpy